test(sm-ui): turn on coverage reports and rename __mockData__ to __mocks__
