+++ packages/slice-machine/tests/__mocks__/widgets/Color.js
