+++ packages/slice-machine/tests/__mocks__/widgets/GeoPoint.js
