+++ packages/slice-machine/tests/__mocks__/widgets/Image.js
