+++ packages/slice-machine/tests/__mocks__/widgets/LinkToMedia.js
