+++ packages/slice-machine/tests/__mocks__/widgets/StructuredText.js
