+++ packages/slice-machine/tests/__mocks__/widgets/Text.js
