+++ packages/slice-machine/tests/__mocks__/widgets/Timestamp.js
